move ui to main_ui
